requirements.txt
django
djangorestframework
requests
urllib3
behave
discord
